requirements.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
docopt==0.6.2
idna==2.10
pipreqs==0.4.10
praw==7.1.0
prawcore==1.4.0
pycparser==2.20
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
update-checker==0.17
urllib3==1.25.10
websocket-client==0.57.0
yarg==0.1.9
